Don't run git clean on CI

+++ Rakefile
@@ -153,7 +153,7 @@ end
 # Tasks
 #
 task :clean do
-  sh "git clean -fdx"
+  sh "git clean -fdx" unless ci?
 end
 
 namespace :package do
